requirements.txt
streamlit
Pillow
easyocr
hashlib
